OPTIMISATION: import des réglages de chaque tableau depuis le dossier options
