+++ frontend/src/pages/ChartsCO2.jsx
@@ -1,5 +1,4 @@
 import { useState, useEffect } from "react";
-import options from "../options/options_CO2";
 import { Line } from "react-chartjs-2";
 import {
   Chart as ChartJS,
@@ -12,6 +11,8 @@ import {
   Legend,
 } from "chart.js";
 
+import reglages from "../options/options_CO2";
+
 ChartJS.register(
   CategoryScale,
   LinearScale,
@@ -21,45 +22,30 @@ ChartJS.register(
   Tooltip,
   Legend
 );
-// constantes de données, api distante et copie locale
-const apiOnline = "https://global-warming.org/api/co2-api";
-const apiLocale = "./datasets/co2.json";
+
+/**
+ *  initialisation des constantes importées
+ */
+// const apiOnline = reglages.apiOnline;
+// const apiLocale = reglages.apiLocale;
+// const options = reglages.options;
+// const dataModel = reglages.dataModel;
 const CO2 = "CO\u2082";
 // fin des constantes de données
 
-/**
- *  réglages des axes du tableau
- */
-const dataModel = {
-  labels: [],
-  datasets: [
-    {
-      label: "Valeurs quotidiennes (cycle)", // titre du premier graphique
-      data: [],
-      pointStyle: "dash",
-      radius: 0.1,
-      borderColor: "rgba(136, 78, 160,0.8)",
-      backgroundColor: "rgba(136, 78, 160, 0.8)",
-    },
-    {
-      label: "Tendance moyenne (évolution)", // titre du second graphique
-      data: [],
-      pointStyle: "dash",
-      borderColor: "rgba(52,73,94, 0.8)", // "rgb(93, 109, 126)",
-      backgroundColor: "rgba(52, 73, 94, 0.8)",
-    },
-  ],
-};
-
 function ChartsCO2() {
-  const [dataSet, setDataSet] = useState(dataModel);
-
+  const [dataSet, setDataSet] = useState(reglages.dataModel);
   const prepaConfig = (data) => {
-    options.plugins.title.text = `Concentration de Dioxyde de Carbone dans l'atmosphère (de ${
+    // légendage
+    reglages.options.plugins.title.text = `Concentration de Dioxyde de Carbone dans l'atmosphère (de ${
       data[0].year
     } à ${data[data.length - 1].year})`;
   };
 
+  /**
+   * préparation des données de l'API
+   * @param {array} data
+   */
   const prepareDataSet = (data) => {
     /** Récupération du jeu de dates */
     const dataSetProv = { ...dataSet };
@@ -71,6 +57,12 @@ function ChartsCO2() {
     setDataSet(dataSetProv);
   };
 
+  /**
+   * fetch des données de l'API
+   * En cas d'erreur, lecture du json local
+   * @param {api} url
+   * @param {function} callback
+   */
   const getStaticData = (url, callback) => {
     fetch(url)
       .then((res) => res.json())
@@ -79,16 +71,20 @@ function ChartsCO2() {
         prepaConfig(data.co2);
       })
       // ternaire de fetch des données locales en cas d'erreur
-      .catch((err) => (callback ? callback(apiLocale) : console.error(err)));
+      .catch((err) =>
+        callback ? callback(reglages.apiLocale) : console.error(err)
+      );
   };
 
   useEffect(() => {
-    getStaticData(apiOnline, getStaticData);
+    getStaticData(reglages.apiOnline, getStaticData);
   }, []);
   return (
     // affichage du composant graphique
     <div className="graph_render_dd">
-      {dataSet.labels.length > 0 && <Line data={dataSet} options={options} />}
+      {dataSet.labels.length > 0 && (
+        <Line data={dataSet} options={reglages.options} />
+      )}
       <p className="CetteDivDoitDisparaitre_dd">parlons {CO2}</p>
     </div>
   );
